test: add missing board generator test
