--- packages/board_generator/test/src/board_generator_test.dart
@@ -1,8 +1,0 @@
-// ignore_for_file: prefer_const_constructors
-import 'package:test/test.dart';
-
-void main() {
-  group('BoardGenerator', () {
-    test('can be instantiated', () {});
-  });
-}
